test_az_func/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
altair==4.2.2
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.5
bqplot==0.12.36
branca==0.6.0
bs4==0.0.1
cachetools==5.3.0
cdsapi==0.5.1
certifi==2022.12.7
cffi==1.15.1
cftime==1.6.2
charset-normalizer==3.0.1
click==8.1.3
colour==0.1.5
comm==0.1.2
constantly==15.1.0
contourpy==1.0.7
cryptography==39.0.1
cssselect==1.2.0
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
eccodes==1.5.1
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
fake-useragent==1.1.3
fastjsonschema==2.16.3
filelock==3.9.0
findlibs==0.0.2
folium==0.13.0
fonttools==4.38.0
fqdn==1.5.1
gdown==4.6.4
geojson==3.0.1
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
incremental==22.10.0
install==1.3.5
ipyevents==2.0.1
ipyfilechooser==0.6.0
ipykernel==6.21.3
ipyleaflet==0.17.2
ipython==8.11.0
ipython-genutils==0.2.0
webdriver_manager
ipytree==0.2.2
ipywidgets==7.7.3
isoduration==20.11.0
itemadapter==0.7.0
itemloaders==1.0.6
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.4.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.2
kiwisolver==1.4.4
leafmap==0.18.2
lxml==4.9.2
Magics==1.5.8
magpye==0.1.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdurl==0.1.2
metview==1.13.1
mistune==2.0.5
multiurl==0.2.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
netCDF4==1.6.2
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parse==1.19.0
parsel==1.7.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.0
plotly==5.13.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
Protego==0.2.1
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0
PyDispatcher==2.0.7
pyee==8.2.2
Pygments==2.14.0
Pympler==1.0.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==2.0.0
pyrsistent==0.19.3
pyshp==2.3.1
PySocks==1.7.1
pystac==1.6.1
pystac-client==0.6.0
python-box==7.0.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.0
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
requests-html==0.10.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.2
scipy==1.10.1
scooby==0.7.1
Scrapy==2.8.0
selenium==4.8.3
semver==2.13.0
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
stack-data==0.6.2
streamlit==1.19.0
tenacity==8.2.2
terminado==0.17.1
tinycss2==1.2.1
tldextract==3.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
traittypes==0.2.1
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.10.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uri-template==1.2.0
urllib3==1.26.14
validators==0.20.0
w3lib==2.1.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
wget==3.2
whitebox==2.2.0
whiteboxgui==2.2.0
widgetsnbextension==3.6.2
wsproto==1.2.0
xarray==2023.2.0
xyzservices==2023.2.0
zipp==3.14.0
zope.interface==5.5.2
webdriver_manager
